+++ src/message.c
@@ -126,6 +126,7 @@ extern struct bar g_bar;
 #define ARGUMENT_WINDOW_GRID          "%d:%d:%d:%d:%d:%d"
 #define ARGUMENT_WINDOW_MOVE          "%255[^:]:%f:%f"
 #define ARGUMENT_WINDOW_RESIZE        "%255[^:]:%f:%f"
+#define ARGUMENT_WINDOW_TOGGLE_ON_TOP "topmost"
 #define ARGUMENT_WINDOW_TOGGLE_FLOAT  "float"
 #define ARGUMENT_WINDOW_TOGGLE_STICKY "sticky"
 #define ARGUMENT_WINDOW_TOGGLE_SHADOW "shadow"
@@ -158,6 +159,7 @@ extern struct bar g_bar;
 #define ARGUMENT_RULE_KEY_ALPHA   "opacity"
 #define ARGUMENT_RULE_KEY_MANAGE  "manage"
 #define ARGUMENT_RULE_KEY_STICKY  "sticky"
+#define ARGUMENT_RULE_KEY_ON_TOP  "topmost"
 #define ARGUMENT_RULE_KEY_BORDER  "border"
 #define ARGUMENT_RULE_KEY_FULLSCR "native-fullscreen"
 #define ARGUMENT_RULE_KEY_GRID    "grid"
@@ -368,7 +370,7 @@ static void handle_domain_config(FILE *rsp, struct token domain, char *message)
         } else {
             daemon_fail(rsp, "unknown value '%.*s' given to command '%.*s' for domain '%.*s'\n", value.length, value.text, command.length, command.text, domain.length, domain.text);
         }
-    } else if (token_equals(command, COMMAND_CONFIG_OPACITY)) {
+    } else if (token_equals(command, COMMAND_CONFIG_OPACITY_DURATION)) {
         struct token value = get_token(&message);
         if (!token_is_valid(value)) {
             fprintf(rsp, "%f\n", g_window_manager.window_opacity_duration);
@@ -1199,12 +1201,28 @@ static void handle_domain_space(FILE *rsp, struct token domain, char *message)
     } else if (token_equals(command, COMMAND_SPACE_DISPLAY)) {
         struct selector selector = parse_display_selector(rsp, &message, display_manager_active_display_id());
         if (selector.did_parse && selector.did) {
-            space_manager_move_space_to_display(&g_space_manager, acting_sid, selector.did);
+            enum space_op_error result = space_manager_move_space_to_display(&g_space_manager, acting_sid, selector.did);
+            if (result == SPACE_OP_ERROR_MISSING_SRC) {
+                daemon_fail(rsp, "could not locate the space to act on.\n");
+            } else if (result == SPACE_OP_ERROR_MISSING_DST) {
+                daemon_fail(rsp, "could not locate the active space of the given display.\n");
+            } else if (result == SPACE_OP_ERROR_INVALID_SRC) {
+                daemon_fail(rsp, "acting space is the last user-space on the source display and cannot be moved.\n");
+            } else if (result == SPACE_OP_ERROR_INVALID_DST) {
+                daemon_fail(rsp, "acting space is already located on the given display.\n");
+            }
         }
     } else if (token_equals(command, COMMAND_SPACE_CREATE)) {
         space_manager_add_space(acting_sid);
     } else if (token_equals(command, COMMAND_SPACE_DESTROY)) {
-        space_manager_destroy_space(acting_sid);
+        enum space_op_error result = space_manager_destroy_space(acting_sid);
+        if (result == SPACE_OP_ERROR_MISSING_SRC) {
+            daemon_fail(rsp, "could not locate the space to act on.\n");
+        } else if (result == SPACE_OP_ERROR_INVALID_SRC) {
+            daemon_fail(rsp, "acting space is the last user-space on the source display and cannot be destroyed.\n");
+        } else if (result == SPACE_OP_ERROR_INVALID_TYPE) {
+            daemon_fail(rsp, "cannot destroy a macOS fullscreen space.\n");
+        }
     } else if (token_equals(command, COMMAND_SPACE_BALANCE)) {
         space_manager_balance_space(&g_space_manager, acting_sid);
     } else if (token_equals(command, COMMAND_SPACE_MIRROR)) {
@@ -1353,6 +1371,8 @@ static void handle_domain_window(FILE *rsp, struct token domain, char *message)
         struct token value = get_token(&message);
         if (token_equals(value, ARGUMENT_WINDOW_TOGGLE_FLOAT)) {
             window_manager_toggle_window_float(&g_space_manager, &g_window_manager, acting_window);
+        } else if (token_equals(value, ARGUMENT_WINDOW_TOGGLE_ON_TOP)) {
+            window_manager_toggle_window_topmost(acting_window);
         } else if (token_equals(value, ARGUMENT_WINDOW_TOGGLE_STICKY)) {
             window_manager_toggle_window_sticky(&g_space_manager, &g_window_manager, acting_window);
         } else if (token_equals(value, ARGUMENT_WINDOW_TOGGLE_SHADOW)) {
@@ -1616,6 +1636,12 @@ static void handle_domain_rule(FILE *rsp, struct token domain, char *message)
                     rule->sticky = RULE_PROP_ON;
                 } else if (string_equals(value, ARGUMENT_COMMON_VAL_OFF)) {
                     rule->sticky = RULE_PROP_OFF;
+                }
+            } else if (string_equals(key, ARGUMENT_RULE_KEY_ON_TOP)) {
+                if (string_equals(value, ARGUMENT_COMMON_VAL_ON)) {
+                    rule->topmost = RULE_PROP_ON;
+                } else if (string_equals(value, ARGUMENT_COMMON_VAL_OFF)) {
+                    rule->topmost = RULE_PROP_OFF;
                 }
             } else if (string_equals(key, ARGUMENT_RULE_KEY_BORDER)) {
                 if (string_equals(value, ARGUMENT_COMMON_VAL_ON)) {